data_crawling/02/requirements/requirements.dev.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.6
Scrapy==1.7.2
gunicorn==19.9.0
python-dotenv==0.10.3
psycopg2==2.8.3